os: Split into os/{memory,system_info,xdg}
