+++ os/system_info.h
@@ -1,0 +1,12 @@
+// SPDX-FileCopyrightText: 2022-2023 Robin Lindén <dev@robinlinden.eu>
+//
+// SPDX-License-Identifier: BSD-2-Clause
+
+#ifndef OS_SYSTEM_INFO_H_
+#define OS_SYSTEM_INFO_H_
+
+namespace os {
+unsigned active_window_scale_factor();
+} // namespace os
+
+#endif
